# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.AvatarResponse:
    properties:
      avatar_url:
        example: /uploads/avatars/users/458622d8-daba-4252-8ce1-846277353139.png
        type: string
      message:
        example: Avatar uploaded
        type: string
    type: object
  dto.ContactEnvelope:
    properties:
      contact:
        $ref: '#/definitions/models.Contact'
      message:
        type: string
    type: object
  dto.ContactsEnvelope:
    properties:
      contacts:
        items:
          $ref: '#/definitions/models.Contact'
        type: array
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.PageMeta'
    type: object
  dto.CreateContactRequest:
    properties:
      name:
        example: Work email
        maxLength: 100
        type: string
      org_id:
        description: |-
          OrgID attaches the contact to a team; nil means personal.
          HTTP handlers overwrite UserID with the caller (owners create only
          their own contacts) and check org membership when set.
        example: b3c4d5e6-f7a8-49b0-c1d2-e3f4a5b6c7d8
        type: string
      type:
        enum:
        - email
        - phone
        - address
        - other
        example: email
        type: string
      user_id:
        description: |-
          UserID is accepted for compatibility but ignored by HTTP handlers,
          which always force ownership to the caller.
        example: 458622d8-daba-4252-8ce1-846277353139
        maxLength: 36
        type: string
      value:
        example: kaige@work.com
        maxLength: 2048
        type: string
    required:
    - type
    - value
    type: object
...
  dto.ErrorEnvelope:
    properties:
      error:
        type: string
    type: object
...
  dto.ImportSummary:
    properties:
      errors:
        items:
          $ref: '#/definitions/dto.RowError'
        type: array
      failed:
        example: 1
        type: integer
      imported:
        example: 2
        type: integer
      message:
        example: Import finished
        type: string
    type: object
...
  dto.MessageEnvelope:
    properties:
      message:
        type: string
    type: object
...
  dto.PageMeta:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  dto.RowError:
    properties:
      message:
        example: invalid email value
        type: string
      row:
        example: 3
        type: integer
    type: object
...
  dto.UpdateContactRequest:
    properties:
      name:
        example: Work email
        maxLength: 100
        type: string
      type:
        enum:
        - email
        - phone
        - address
        - other
        example: phone
        type: string
      value:
        example: "+15551234567"
        maxLength: 2048
        type: string
    required:
    - type
    - value
    type: object
...
  models.Contact:
    properties:
      avatar_url:
        example: /uploads/avatars/contacts/9c9e6679-7425-40de-944b-e07fc1f90ae7.png
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 9c9e6679-7425-40de-944b-e07fc1f90ae7
        type: string
      name:
        example: Work email
        type: string
      org_id:
        description: OrgID is nil for personal contacts (owner-only). Set for team
          contacts.
        example: b3c4d5e6-f7a8-49b0-c1d2-e3f4a5b6c7d8
        type: string
      type:
        example: email
        type: string
      updated_at:
        type: string
      user_id:
        example: 458622d8-daba-4252-8ce1-846277353139
        type: string
      value:
        example: kaige@work.com
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Grip API with JWT auth and email verification.
  title: Grip API
  version: "1.0"
paths:
...
  /contacts/:
    get:
      parameters:
      - description: Filter by user ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: query
        name: user_id
        type: string
      - description: Filter by type
        enum:
        - email
        - phone
        - address
        - other
        example: email
        in: query
        name: type
        type: string
      - description: Search name and value
        example: kaige
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        example: 10
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ContactsEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List contacts with pagination and filters
      tags:
      - contacts
    post:
      consumes:
      - application/json
      parameters:
      - description: Contact payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateContactRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ContactEnvelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create contact
      tags:
      - contacts
  /contacts/{id}:
    delete:
      parameters:
      - description: Contact ID
        example: 9c9e6679-7425-40de-944b-e07fc1f90ae7
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Delete contact
      tags:
      - contacts
    get:
      parameters:
      - description: Contact ID
        example: 9c9e6679-7425-40de-944b-e07fc1f90ae7
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Get contact by ID
      tags:
      - contacts
    put:
      consumes:
      - application/json
      parameters:
      - description: Contact ID
        example: 9c9e6679-7425-40de-944b-e07fc1f90ae7
        in: path
        name: id
        required: true
        type: string
      - description: Contact payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateContactRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Contact'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Update contact
      tags:
      - contacts
  /contacts/{id}/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Contact ID
        example: 9c9e6679-7425-40de-944b-e07fc1f90ae7
        in: path
        name: id
        required: true
        type: string
      - description: Avatar image (.jpg, .jpeg, .png, .gif, .webp; max 5 MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvatarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Upload a contact's avatar image
      tags:
      - contacts
  /contacts/import/csv:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Owner user ID
        in: formData
        name: user_id
        required: true
        type: string
      - description: CSV file with header name,type,value
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Import contacts from a CSV file
      tags:
      - contacts
  /contacts/import/json:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Owner user ID
        in: formData
        name: user_id
        required: true
        type: string
      - description: JSON file with an array of {name,type,value}
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImportSummary'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Import contacts from a JSON file
      tags:
      - contacts
...
  /users/{id}/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: User ID
        example: 458622d8-daba-4252-8ce1-846277353139
        in: path
        name: id
        required: true
        type: string
      - description: Avatar image (.jpg, .jpeg, .png, .gif, .webp; max 5 MB)
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AvatarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Upload a user's avatar image
      tags:
      - users
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT. Example: "Bearer
      eyJhbGciOiJIUzI1NiJ9..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"